Changed icon and fix lint issue
